+++ osu.Game/Overlays/OptionsOverlay.cs
@@ -120,99 +120,4 @@ namespace osu.Game.Overlays
             MoveToX(-width, 300, EasingTypes.Out);
         }
     }
-
-    public class OptionsSection : Container
-    {
-        private SpriteText header;
-        private FlowContainer content;
-        protected override Container Content => content;
-        
-        public string Header
-        {
-            get { return header.Text; }
-            set { header.Text = value; }
-        }
-        
-        public OptionsSection()
-        {
-            const int headerSize = 30, headerMargin = 25;
-            const int borderSize = 2;
-            AutoSizeAxes = Axes.Y;
-            RelativeSizeAxes = Axes.X;
-            AddInternal(new Drawable[]
-            {
-                new Box
-                {
-                    Colour = new Color4(3, 3, 3, 255),
-                    RelativeSizeAxes = Axes.X,
-                    Height = borderSize,
-                },
-                new Container
-                {
-                    Padding = new MarginPadding
-                    {
-                        Top = 10 + borderSize,
-                        Left = OptionsOverlay.SideMargins,
-                        Right = OptionsOverlay.SideMargins,
-                        Bottom = 10,
-                    },
-                    RelativeSizeAxes = Axes.X,
-                    AutoSizeAxes = Axes.Y,
-                    Children = new[]
-                    {
-                        header = new SpriteText
-                        {
-                            TextSize = headerSize,
-                            Colour = new Color4(247, 198, 35, 255),
-                        },
-                        content = new FlowContainer
-                        {
-                            Margin = new MarginPadding { Top = headerSize + headerMargin },
-                            Direction = FlowDirection.VerticalOnly,
-                            Spacing = new Vector2(0, 25),
-                            AutoSizeAxes = Axes.Y,
-                            RelativeSizeAxes = Axes.X,
-                        },
-                    }
-                },
-            });
-        }
-    }
-
-    public class OptionsSubsection : Container
-    {
-        private SpriteText header;
-        private Container content;
-        protected override Container Content => content;
-        
-        public string Header
-        {
-            get { return header.Text; }
-            set { header.Text = value.ToUpper(); }
-        }
-    
-        public OptionsSubsection()
-        {
-            RelativeSizeAxes = Axes.X;
-            AutoSizeAxes = Axes.Y;
-            AddInternal(new Drawable[]
-            {
-                content = new FlowContainer
-                {
-                    Direction = FlowDirection.VerticalOnly,
-                    RelativeSizeAxes = Axes.X,
-                    AutoSizeAxes = Axes.Y,
-                    Spacing = new Vector2(0, 5),
-                    Children = new[]
-                    {
-                        header = new SpriteText
-                        {
-                            TextSize = 25,
-                            // TODO: Bold
-                        }
-                    }
-                },
-            });
-        }
-    }
 }
